Moved table.hpp.
