+++ cpp/benchmarks/type_dispatcher/type_dispatcher_benchmark.cu
@@ -16,7 +16,7 @@
 
 
 #include <cudf/copying.hpp>
-#include <cudf/table.hpp>
+#include <cudf/legacy/table.hpp>
 
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
