+++ cpp/include/cudf/legacy/table.hpp
@@ -17,8 +17,8 @@
 #ifndef TABLE_HPP
 #define TABLE_HPP
 
-#include "cudf.h"
-#include "types.hpp"
+#include <cudf/cudf.h>
+#include <cudf/types.hpp>
 
 #include <cassert>
 #include <initializer_list>
