+++ cpp/src/copying/copy.cpp
@@ -19,7 +19,7 @@
 #include <utilities/column_utils.hpp>
 #include <utilities/error_utils.hpp>
 #include <cudf/cudf.h>
-#include <cudf/table.hpp>
+#include <cudf/legacy/table.hpp>
 #include <nvstrings/NVCategory.h>
 
 #include <cuda_runtime.h>
