+++ cpp/src/copying/scatter.cu
@@ -20,7 +20,7 @@
 #include <rmm/thrust_rmm_allocator.h>
 #include <utilities/cudf_utils.h>
 #include <utilities/type_dispatcher.hpp>
-#include <cudf/table.hpp>
+#include <cudf/legacy/table.hpp>
 
 #include <algorithm>
 #include <thrust/scatter.h>
