+++ cpp/src/merge/merge.cu
@@ -15,7 +15,7 @@
 #include <cudf/cudf.h>
 #include <cudf/types.hpp>
 #include <cudf/copying.hpp>
-#include <cudf/table.hpp>
+#include <cudf/legacy/table.hpp>
 #include <table/legacy/device_table.cuh>
 #include <table/legacy/device_table_row_operators.cuh>
 #include "bitmask/legacy/bit_mask.cuh"
