+++ cpp/src/table/legacy/device_table.cuh
@@ -24,7 +24,7 @@
 #include <bitmask/legacy/legacy_bitmask.hpp>
 #include <hash/hash_functions.cuh>
 #include <hash/managed.cuh>
-#include <cudf/table.hpp>
+#include <cudf/legacy/table.hpp>
 #include <utilities/error_utils.hpp>
 #include <utilities/type_dispatcher.hpp>
