+++ cpp/src/table/legacy/table.cpp
@@ -18,7 +18,7 @@
 #include <bitmask/legacy/legacy_bitmask.hpp>
 #include <cassert>
 #include <cudf/copying.hpp>
-#include <cudf/table.hpp>
+#include <cudf/legacy/table.hpp>
 #include <utilities/column_utils.hpp>
 #include <utilities/error_utils.hpp>
