+++ cpp/tests/copying/gather_tests.cu
@@ -15,7 +15,7 @@
  */
 
 #include <cudf/copying.hpp>
-#include <cudf/table.hpp>
+#include <cudf/legacy/table.hpp>
 
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
