+++ cpp/tests/groupby/hash/single_column_min.cu
@@ -16,7 +16,7 @@
 
 #include <tests/utilities/cudf_test_fixtures.h>
 #include <cudf/groupby.hpp>
-#include <cudf/table.hpp>
+#include <cudf/legacy/table.hpp>
 #include <tests/utilities/column_wrapper.cuh>
 #include <tests/utilities/compare_column_wrappers.cuh>
 #include <utilities/type_dispatcher.hpp>
