+++ cpp/tests/groupby/without_agg/groupby_wo_agg_test.cu
@@ -27,7 +27,7 @@
 
 #include <cudf/cudf.h>
 #include <cudf/groupby.hpp>
-#include <cudf/table.hpp>
+#include <cudf/legacy/table.hpp>
 
 #include <utilities/cudf_utils.h>
